Feature/fill

+++ reservation-service/handlers/reservation.go
@@ -3,6 +3,10 @@ package handler
 import (
 	"context"
 	"fmt"
+	"github.com/dgrijalva/jwt-go"
+	grpc_auth "github.com/grpc-ecosystem/go-grpc-middleware/auth"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 	"res_init/model"
 	"res_init/proto/reservation"
 	"res_init/service"
@@ -13,6 +17,45 @@ func NewReservationHandler(service *service.ResService) *ReservationHandler {
 	return &ReservationHandler{
 		ResService: service,
 	}
+}
+func parseToken(token string) (jwt.MapClaims, error) {
+	parsedToken, err := jwt.Parse(token, func(token *jwt.Token) (interface{}, error) {
+		return []byte("secret_key"), nil
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	claims, ok := parsedToken.Claims.(jwt.MapClaims)
+	if !ok || !parsedToken.Valid {
+		return nil, fmt.Errorf("invalid token")
+	}
+
+	return claims, nil
+}
+
+func userClaimFromToken(claims jwt.MapClaims) string {
+
+	sub, ok := claims["role"].(string)
+	if !ok {
+		return ""
+	}
+
+	return sub
+}
+
+func checkRole(ctx context.Context) string {
+	token, err := grpc_auth.AuthFromMD(ctx, "bearer")
+	if err != nil {
+		return ""
+	}
+
+	tokenInfo, _ := parseToken(token)
+
+	role := userClaimFromToken(tokenInfo)
+
+	fmt.Println("role is: " + role)
+	return role
 }
 
 type ReservationHandler struct {
@@ -26,30 +69,49 @@ func (h ReservationHandler) GreetFromReservation(ctx context.Context, request *r
 	}, nil
 }
 func (h ReservationHandler) Reserve(ctx context.Context, request *reservation.RequestForReserve) (*reservation.ResponseForReserve, error) {
-	fmt.Println(request)
-	var Reservation = model.Reservation{}
-	Reservation.GuestNumber = int(request.GetReserve().GuestNumber)
-	Reservation.Accepted = "0"
-	Reservation.Accommodation = request.GetReserve().Accommodation
-	layout := "2006-01-02T15:04:05Z"
-	Reservation.FromDate, _ = time.Parse(layout, request.GetReserve().FromDate)
-	Reservation.ToDate, _ = time.Parse(layout, request.GetReserve().ToDate)
-	//objectId, _ := primitive.ObjectIDFromHex(accoId)
-	h.ResService.Create(&Reservation)
+	role := checkRole(ctx)
+	if role == "User" {
+		fmt.Println(request)
+		var Reservation = model.Reservation{}
+		Reservation.GuestNumber = int(request.GetReserve().GuestNumber)
+		Reservation.Accepted = "0"
+		Reservation.Accommodation = request.GetReserve().Accommodation
+		layout := "2006-01-02T15:04:05Z"
+		Reservation.FromDate, _ = time.Parse(layout, request.GetReserve().FromDate)
+		Reservation.ToDate, _ = time.Parse(layout, request.GetReserve().ToDate)
+		//objectId, _ := primitive.ObjectIDFromHex(accoId)
+		h.ResService.Create(&Reservation)
+		return &reservation.ResponseForReserve{
+			Reserve: fmt.Sprintf("Succesfully created! %s", request),
+		}, nil
+	}
 	return &reservation.ResponseForReserve{
-		Reserve: fmt.Sprintf("Succesfully created! %s", request),
-	}, nil
+		Reserve: "Error",
+	}, status.Errorf(codes.Unauthenticated, "You don't have permissions for this action")
 }
 func (h ReservationHandler) DeleteReservation(ctx context.Context, request *reservation.RequestDeleteReservation) (*reservation.ResponseDeleteReservation, error) {
-	h.ResService.DeleteReservation(request.Delres)
+	role := checkRole(ctx)
+	if role == "User" || role == "Host" {
+		h.ResService.DeleteReservation(request.Delres)
+		return &reservation.ResponseDeleteReservation{
+			Delres: "Deleted",
+		}, nil
+	}
 	return &reservation.ResponseDeleteReservation{
-		Delres: "Deleted",
-	}, nil
+		Delres: "Error",
+	}, status.Errorf(codes.Unauthenticated, "You don't have permissions for this action")
+
 }
 func (h ReservationHandler) AcceptReservation(ctx context.Context, request *reservation.DeleteRequest) (*reservation.DeleteResponse, error) {
-	fmt.Println(request.Dlt + " acce")
-	h.ResService.AcceptReservation(request.Dlt)
+	role := checkRole(ctx)
+	if role == "Host" {
+		fmt.Println(request.Dlt + " acce")
+		h.ResService.AcceptReservation(request.Dlt)
+		return &reservation.DeleteResponse{
+			Dlt: "Updated",
+		}, nil
+	}
 	return &reservation.DeleteResponse{
-		Dlt: "Updated",
-	}, nil
+		Dlt: "Error",
+	}, status.Errorf(codes.Unauthenticated, "You don't have permissions for this action")
 }
